requirements.txt
Flask==2.3.2
line-bot-sdk==3.2.0
requests==2.31.0